change do_idea to component
